reporter: Move most logic of `NoticeReporter` to an abstract class
Only keep the logic that adds the licenses to the notice file in the
implementation of an abstract function in `NoticeReporter`. This allows
to add another notice reporter using a different algorithm for adding
licenses in the next commit.
